+++ seatunnel-connectors-v2/connector-iceberg/src/test/java/org/apache/seatunnel/connectors/seatunnel/iceberg/utils/ExpressionUtilsTest.java
@@ -30,6 +30,8 @@ import net.sf.jsqlparser.parser.CCJSqlParserUtil;
 import net.sf.jsqlparser.statement.Statement;
 import net.sf.jsqlparser.statement.delete.Delete;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 public class ExpressionUtilsTest {
 
     @Test
@@ -127,5 +129,125 @@ public class ExpressionUtilsTest {
                                 Expressions.equal("f3", 1704110400001000L))
                         .toString(),
                 expression.toString());
+    }
+
+    @Test
+    public void testSimpleConditions() throws Exception {
+        // test integer comparison
+        String whereClause1 = "age = 30";
+        Expression expr1 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause1);
+        assertEquals(Expressions.equal("age", 30).toString(), expr1.toString());
+
+        // test string comparison
+        String whereClause2 = "name = 'John'";
+        Expression expr2 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause2);
+        assertEquals(Expressions.equal("name", "John").toString(), expr2.toString());
+
+        // test float comparison
+        String whereClause3 = "salary > 50000.5";
+        Expression expr3 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause3);
+        assertEquals(Expressions.greaterThan("salary", 50000.5).toString(), expr3.toString());
+
+        // test boolean comparison
+        String whereClause4 = "is_active is true";
+        Expression expr4 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause4);
+        assertEquals(Expressions.equal("is_active", true).toString(), expr4.toString());
+    }
+
+    @Test
+    public void testLogicalCombinations() throws Exception {
+        // test AND
+        String whereClause1 = "age > 30 AND name = 'John'";
+        Expression expr1 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause1);
+        assertEquals(
+                Expressions.and(
+                                Expressions.greaterThan("age", 30),
+                                Expressions.equal("name", "John"))
+                        .toString(),
+                expr1.toString());
+
+        // OR
+        String whereClause2 = "salary < 50000 OR is_active is true";
+        Expression expr2 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause2);
+        assertEquals(
+                Expressions.or(
+                                Expressions.lessThan("salary", 50000),
+                                Expressions.equal("is_active", true))
+                        .toString(),
+                expr2.toString());
+
+        // test combination of AND and OR
+        String whereClause3 = "(age > 30 AND name = 'John') OR salary < 50000";
+        Expression expr3 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause3);
+        assertEquals(
+                Expressions.or(
+                                Expressions.and(
+                                        Expressions.greaterThan("age", 30),
+                                        Expressions.equal("name", "John")),
+                                Expressions.lessThan("salary", 50000))
+                        .toString(),
+                expr3.toString());
+    }
+
+    @Test
+    public void testComplexNestedExpressions() throws Exception {
+        // test nested AND and OR
+        String whereClause1 =
+                "((age > 30 AND name = 'John') OR salary < 50000) AND is_active is true";
+        Expression expr1 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause1);
+        assertEquals(
+                Expressions.and(
+                                Expressions.or(
+                                        Expressions.and(
+                                                Expressions.greaterThan("age", 30),
+                                                Expressions.equal("name", "John")),
+                                        Expressions.lessThan("salary", 50000)),
+                                Expressions.equal("is_active", true))
+                        .toString(),
+                expr1.toString());
+
+        // test nested AND and OR with multiple levels
+        String whereClause2 =
+                "age > 30 AND (name = 'John' OR (salary < 50000 AND is_active is true))";
+        Expression expr2 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause2);
+        assertEquals(
+                Expressions.and(
+                                Expressions.greaterThan("age", 30),
+                                Expressions.or(
+                                        Expressions.equal("name", "John"),
+                                        Expressions.and(
+                                                Expressions.lessThan("salary", 50000),
+                                                Expressions.equal("is_active", true))))
+                        .toString(),
+                expr2.toString());
+    }
+
+    @Test
+    public void testSpecialScenarios() throws Exception {
+        // IS NULL
+        String whereClause1 = "name IS NULL";
+        Expression expr1 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause1);
+        assertEquals(Expressions.isNull("name").toString(), expr1.toString());
+
+        // IS NOT NULL
+        String whereClause2 = "name IS NOT NULL";
+        Expression expr2 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause2);
+        assertEquals(Expressions.notNull("name").toString(), expr2.toString());
+
+        // NOT
+        String whereClause3 = "NOT (age > 30)";
+        Expression expr3 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause3);
+        assertEquals(
+                Expressions.not(Expressions.greaterThan("age", 30)).toString(), expr3.toString());
+
+        // IN
+        String whereClause4 = "age IN (30, 40, 50)";
+        Expression expr4 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause4);
+        assertEquals(Expressions.in("age", new Object[] {30, 40, 50}).toString(), expr4.toString());
+
+        // start with
+        String whereClause5 = "name LIKE 'John%'";
+        Expression expr5 = ExpressionUtils.parseWhereClauseToIcebergExpression(whereClause5);
+        assertEquals(Expressions.startsWith("name", "John%").toString(), expr5.toString());
     }
 }
